Add support for major IPA consonants
Most of the consonants on the IPA chart 2020
edition are now added to the IpaRegistry class.

Support for previewing the such IPA consonants is
also added.

Structures now support basic preview for the
output string and Formulang transcript.

Articulatory properties are now represented by
bonafide classes instead of enum classes.